Рефакторинг модуля Service

+++ commons/src/main/java/ru/cristalix/museum/packages/ThanksExecutePackage.java
@@ -15,6 +15,6 @@ public class ThanksExecutePackage extends MuseumPackage {
 	private UUID user;
 
 	// response
-	private int boostersCount; // Количество бустеров, за которые поблагодарил игрок
+	private long boostersCount; // Количество бустеров, за которые поблагодарил игрок
 
 }
